requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
cryptography==44.0.0
pydantic==2.10.5
requests==2.32.3
